sycophancy-sae/requirements.txt
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
datasets>=2.18.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.2.0
pyyaml>=6.0.1
tqdm>=4.66.0
anthropic>=0.25.0
openai>=1.30.0
einops>=0.7.0
# For notebooks
jupyterlab>=4.1.0
